+++ scripts/ppo_gi.py
@@ -39,6 +39,7 @@ from nip.utils.experiments import (
     MultiprocessHyperparameterExperiment,
     SequentialHyperparameterExperiment,
     ExperimentFunctionArguments,
+    RunIDFunctionArguments,
 )
 
 script_name = os.path.basename(__file__)
@@ -348,25 +349,25 @@ def experiment_fn(arguments: ExperimentFunctionArguments):
     )
 
 
-def run_id_fn(combo_index: int | None, cmd_args: Namespace) -> str:
+def run_id_fn(arguments: RunIDFunctionArguments) -> str:
     """Generate the run ID for a given hyperparameter combination.
 
     Parameters
     ----------
-    combo_index : int | None
-        The index of the hyperparameter combination. If None, the run ID is for the
-        entire experiment.
-    cmd_args : Namespace
-        The command line arguments.
+    arguments : RunIDFunctionArguments
+        The arguments for generating the run ID, including:
+
+        - combo_index: The index of the hyperparameter combination.
+        - cmd_args: The command line arguments.
 
     Returns
     -------
     run_id : str
         The run ID.
     """
-    if combo_index is None:
-        return f"ppo_gi_{cmd_args.run_infix}"
-    return f"ppo_gi_{cmd_args.run_infix}_{combo_index}"
+    if arguments.combo_index is None:
+        return f"ppo_gi_{arguments.cmd_args.run_infix}"
+    return f"ppo_gi_{arguments.cmd_args.run_infix}_{arguments.combo_index}"
 
 
 def run_preparer_fn(combo: dict, cmd_args: Namespace) -> PreparedExperimentInfo: